requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
huggingface-hub
langchain-community
langchain-huggingface
pillow
